requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
pandas==2.1.4
python-dotenv==1.0.0
webdriver-manager==4.0.1
multion==0.1.4
agentql==1.11.0 